Separate functions into index.js
